+++ dist_q_learning/experiments/event_experiment/saved_results/init_result.png
